support-queue-assistant/requirements.txt
psycopg2-binary
python-dotenv
uuid
openai
scikit-learn
pandas
flask
streamlit